refactor(ciena_saos): rename kind from ciena_saos10 to ciena_saos
Drop the version from the kind name to keep it version-agnostic,
matching the vrnetlab image rename to vrnetlab/ciena_saos. Kind
aliases become ciena_saos / vr-ciena_saos; Go package, schema
entries, docs page, and lab example are renamed accordingly. The
human-readable "Ciena SAOS 10" display name is unchanged.
