+++ core/register.go
@@ -11,7 +11,7 @@ import (
 	clabnodesc8000 "github.com/srl-labs/containerlab/nodes/c8000"
 	clabnodesceos "github.com/srl-labs/containerlab/nodes/ceos"
 	clabnodescheckpoint_cloudguard "github.com/srl-labs/containerlab/nodes/checkpoint_cloudguard"
-	clabnodesciena_saos10 "github.com/srl-labs/containerlab/nodes/ciena_saos10"
+	clabnodesciena_saos "github.com/srl-labs/containerlab/nodes/ciena_saos"
 	clabnodescisco_asav "github.com/srl-labs/containerlab/nodes/cisco_asav"
 	clabnodescisco_sdwan "github.com/srl-labs/containerlab/nodes/cisco_sdwan"
 	clabnodescisco_vios "github.com/srl-labs/containerlab/nodes/cisco_vios"
@@ -72,7 +72,7 @@ func (c *CLab) RegisterNodes() { //nolint:funlen
 	clabnodesbridge.Register(c.Reg)
 	clabnodesceos.Register(c.Reg)
 	clabnodescheckpoint_cloudguard.Register(c.Reg)
-	clabnodesciena_saos10.Register(c.Reg)
+	clabnodesciena_saos.Register(c.Reg)
 	clabnodescisco_sdwan.Register(c.Reg)
 	clabnodescrpd.Register(c.Reg)
 	clabnodesjuniper_csrx.Register(c.Reg)
