+++ docs/manual/kinds/index.md
@@ -77,7 +77,7 @@ Within each predefined kind, we store the necessary information that is used to 
 | **Palo Alto PAN-OS**       | [`paloalto_panos`](vr-pan.md)                       | supported |    VM     |
 | **Fortinet Fortigate**     | [`fortinet_fortigate`](fortinet_fortigate.md)       | supported |    VM     |
 | **Check Point Cloudguard** | [`checkpoint_cloudguard`](checkpoint_cloudguard.md) | supported |    VM     |
-| **Ciena SAOS 10**          | [`ciena_saos10`](ciena_saos10.md)                   | supported |    VM     |
+| **Ciena SAOS 10**          | [`ciena_saos`](ciena_saos.md)                   | supported |    VM     |
 | **6WIND VSR**              | [`6wind_vsr`](6wind_vsr.md)                         | supported | container |
 | **Keysight ixia-c-one**    | [`keysight_ixia-c-one`](keysight_ixia-c-one.md)     | supported | container |
 | **Spirent TestCenter**     | [`spirent_stc`](spirent_stc.md)                     | supported | container |
