+++ lab-examples/ciena_saos/saos1.cfg.partial
