+++ lab-examples/ciena_saos/saos2.cfg.partial
