+++ schemas/clab.schema.json
@@ -100,7 +100,7 @@
                         "keysight_ixia-c-one",
                         "ipinfusion_ocnos",
                         "checkpoint_cloudguard",
-                        "ciena_saos10",
+                        "ciena_saos",
                         "ext-container",
                         "rare",
                         "cisco_xrd",
@@ -2128,7 +2128,7 @@
                         "checkpoint_cloudguard": {
                             "$ref": "#/definitions/node-config"
                         },
-                        "ciena_saos10": {
+                        "ciena_saos": {
                             "$ref": "#/definitions/node-config"
                         },
                         "ext-container": {
